Updated navbar to change based on window size for logged in and logged out
